webpy/requirements.txt
streamlit
pandas
sklearn.model_selection 
sklearn.preprocessing 
sklearn.ensemble
matplotlib.pyplot 
seaborn 